Add a middleware that allows reuse of try/catch block
The try/catch block were being repeated inside of every single one of
the taskController's handler functions. To comply with the DRY principle,
the asyncWrapper middleware was added.

This middleware accepts a function as an arugment, executes the function
within a try/catch block, and returns a function which gets access to
anything Express passes down to it.

The data it receives from Express (req, res etc.) can be used with the
function it receives as an argument.

This middleware returns any data returned by the function it receives as
an arugment when the function resolves, but it will forward any errors
to the next middleware.